move verbosity setting to verbosity module
