+++ bluecellulab/__init__.py
@@ -9,6 +9,7 @@ except ImportError:
 
 from .importer import *  # NOQA
 from .tools import *  # NOQA
+from .verbosity import *
 from .cell import Cell, create_ball_stick  # NOQA
 from .connection import Connection  # NOQA
 from .plotwindow import PlotWindow  # NOQA
